Add a cache for the token and take an entity ID.
The token is now in a memoised method which caches the token for the
period passed into the client. There is a default of 60 seconds but this
can be overriden by the calling code.

I've also changed the last updated entity methods to return a trait and
the get metadata methods to take a value of the trait. Based on the type
of the entity, the url is constructed in the client. This keeps the urls
out of the calling code.