feat: DR2-1123 Disk based caching.
This now uses a custom class which extends ScalaCache AbstractCache

This uses the file system in the /tmp directory to store values in the
cache. Expiration times are put as attributes on the files themselves
when the item is added to the cache.

I've updated the tests as well. The disk cache persists between tests
now so I need to delete the temporary files it creates