Make entity type optional.
Entities which have been soft deleted in Preservica don't have an entity
type. At the moment, the entity event generator is failing because the
type is missing and an error is being thrown.

It's now optional. If you try to pass this entity to the
metadataForEntity method, it won't work but that's fine.