DR2-1390 Fix invalid XML characters
There are two fixes here.

If you do and if without an else in Scala and the if check is false, the
if statement returns `Unit` This prints as `()` so if the ref or the
description weren't populated, the add entity XML had a line with `()`
in it.
Swapping it out for `.map` and `.getOrElse` is a bit more functional and
avoids this problem.

The other problem is that we're not escaping XML special characters. A
few titles have got `&` in them which needs to be escaped to &amp; The
call to `Utility.escape` for title and description should fix this.