DR2-1543 Remove file system cache.
I've deleted the custom cache and all associated tests.

The in memory cache doesn't need testing as that's third party code.

As a happy side effect, it seems this has fixed the bug where we were
calling the Preservica authentication endpoint twice so I've updated a
couple of the tests to reflect that.