DR2-1703 Remove dynamo formatters dependency.
I want to move the dynamo formatters to the monorepo but I can't while
this is using it.

It shouldn't be using it, the worlds of Preservica and Dynamo are
separate and it's only in there for the Identifier case class. There'll
be some mapping between them in dr2-ingest but I can live with that.