DR2-1257 Add methods to reset and test password changes
I've added a method in to reset the password and one to check we can get
a token with the new password.

I've also replaced the direct calls to AWS secrets manager SDK with the
da-aws-clients one. This meant I had to change all the implicit Sync for
Async which is why there's more changes.