Get URL from secret
This was a bit fiddly because the URL is passed in on client creation.

It now calls secrets manager when the client is created and this is then
cached so we shouldn't be making any more secrets manager calls than we
already were.

I've tested this on the ingest lambdas and custodial copy and it works
with some changes