Update to use Central Publisher Portal
The OSSRH APIs we were using to publish the library are end of life now.

We need to move to using the Maven Central Publisher Portal.

The sbt-sonatype plugin we were using doesn't do this because it's now
built into standard sbt.

I've removed the sbt-sonatype plugin and updated the settings based on
these release notes https://github.com/sbt/sbt/releases/tag/v1.11.0

I've tested this and it's working.