Add retries with exponential backoff
This adds retries to the json and xml requests.

There is a configurable number of retries that defaults to 5 and a hard
coded 1 second initial duration for the backoff.

This will retry on errors, i.e. an exception or on a non 200 response.

We were getting issues in the reconciler where the secret was being
rotated but that cache still had the old secret value which had expired
so we were getting 401/403. We're now checking for this and if it
happens, we invalidate both caches so a new secret value and token can
be picked up.