Remove circe-generic
For some reason I don't understand, any time we pass an implicit decoder
via circe-generic auto derivation, it fails with `NoClassDefFoundError: io/circe/derivation/ConfiguredDecoder$`

If you explicitly create the decoders, it works fine so this is what
I've done here.

It may work with the semi-auto derivation, I haven't tried but this
definitely works so we're going with that.