Include the token in the request.
The response that's passed to the retrier is now getting the token first
and then sending the request.

This way, if we get a 401 because the password has changed, we can
invalidate the cache and the retry will fetch the token again. This
wasn't happening before.

This may cause an extremely tiny number of extra token calls but it's
not a lot as the caching still works.