Fix caching issue.
The cache for the secrets manager call as is isn't working.
It's calling secrets manager endpoints on almost every call.
I suspect it's because there are lines outside an Async block which is
causing something to not work properly but I don't know exactly.

If I run it without this, I get 1000s of secrets manager calls, if I run
it with this, I get one every 15 minutes as expected.