DR2-2419 Add optional end date to updated since.
This will be used for the CC reconciler to make sure we don't pick up
things that come into the OCFL repo after we've got the list of OCFL
files as this leads to false positives.

I've also disabled the timeouts on the http requests as some requests
take a long time if Preservica is busy. There's a slight chance that
this will lock up the process but it seems unlikely.